Struct iri_string::build::Built

pub struct Built<'a, T: ?Sized> { /* private fields */ }

Display-able IRI build result.

The value of this type can generate an IRI using From/Into traits or Display trait.

Security consideration

This can be stringified or directly printed by std::fmt::Display, but note that this Display does not hide the password part. Be careful not to print the value using Display for Built<_> in public context.

Implementations

impl<S: Spec> Built<'_, RiReferenceStr<S>>

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Returns Ok(()) if the IRI is normalizable by the RFC 3986 algorithm.

impl<S: Spec> Built<'_, RiStr<S>>

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Returns Ok(()) if the IRI is normalizable by the RFC 3986 algorithm.

impl<S: Spec> Built<'_, RiAbsoluteStr<S>>

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Returns Ok(()) if the IRI is normalizable by the RFC 3986 algorithm.

impl<S: Spec> Built<'_, RiRelativeStr<S>>

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Returns Ok(()) if the IRI is normalizable by the RFC 3986 algorithm.