Struct iri_string::mask_password::PasswordMasked

pub struct PasswordMasked<'a, T: ?Sized> { /* private fields */ }

A wrapper of an IRI string that masks the non-empty password when Displayed.

This is a retrun type of mask_password method of IRI string types (such as RiStr::mask_password).

Examples

use iri_string::types::UriReferenceStr;

let iri = UriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();
assert_eq!(masked.to_string(), "http://user:@example.com/path?query");

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

Implementations

impl<'a, S: Spec> PasswordMasked<'a, RiReferenceStr<S>>

pub fn replace_password<D>( &self, alt: D, ) -> PasswordReplaced<'a, RiReferenceStr<S>, D>
where D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

pub fn replace_password_with<F, D>( &self, replace: F, ) -> PasswordReplaced<'a, RiReferenceStr<S>, D>
where F: FnOnce(&str) -> D, D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

let replaced = masked
    .replace_password_with(|password| format!("{{{} chars}}", password.len()));
assert_eq!(
    replaced.to_string(),
    "http://user:{8 chars}@example.com/path?query"
);

impl<'a, S: Spec> PasswordMasked<'a, RiStr<S>>

pub fn replace_password<D>(&self, alt: D) -> PasswordReplaced<'a, RiStr<S>, D>
where D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

pub fn replace_password_with<F, D>( &self, replace: F, ) -> PasswordReplaced<'a, RiStr<S>, D>
where F: FnOnce(&str) -> D, D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

let replaced = masked
    .replace_password_with(|password| format!("{{{} chars}}", password.len()));
assert_eq!(
    replaced.to_string(),
    "http://user:{8 chars}@example.com/path?query"
);

impl<'a, S: Spec> PasswordMasked<'a, RiAbsoluteStr<S>>

pub fn replace_password<D>( &self, alt: D, ) -> PasswordReplaced<'a, RiAbsoluteStr<S>, D>
where D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

pub fn replace_password_with<F, D>( &self, replace: F, ) -> PasswordReplaced<'a, RiAbsoluteStr<S>, D>
where F: FnOnce(&str) -> D, D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

let replaced = masked
    .replace_password_with(|password| format!("{{{} chars}}", password.len()));
assert_eq!(
    replaced.to_string(),
    "http://user:{8 chars}@example.com/path?query"
);

impl<'a, S: Spec> PasswordMasked<'a, RiRelativeStr<S>>

pub fn replace_password<D>( &self, alt: D, ) -> PasswordReplaced<'a, RiRelativeStr<S>, D>
where D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

pub fn replace_password_with<F, D>( &self, replace: F, ) -> PasswordReplaced<'a, RiRelativeStr<S>, D>
where F: FnOnce(&str) -> D, D: Display,

Replaces the password with the given arbitrary content.

Note that the result might be invalid as an IRI since arbitrary string can go to the place of the password.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();

let replaced = masked
    .replace_password_with(|password| format!("{{{} chars}}", password.len()));
assert_eq!(
    replaced.to_string(),
    "http://user:{8 chars}@example.com/path?query"
);