Struct iri_string::normalize::Normalized

pub struct Normalized<'a, T: ?Sized> { /* private fields */ }

Normalized OR resolved IRI.

Resolved IRI can be represented by this type. In that case, the result might not be normalized. If you want the IRI resolution result to be normalized, use enable_normalization method.

Implementations

impl<'a, T: ?Sized> Normalized<'a, T>

pub fn enable_normalization(&mut self)

Enables the normalization.

This lets the normalizer apply the case normalization, percent-encoding normalization, and dot segments removal.

pub fn enable_normalization_preserving_authorityless_relative_path(&mut self)

Enables the normalization that preserve relative path under some condition.

Note that this normalization algorithm is not compatible with RFC 3986 algorithm for some inputs.

See RiStr::normalize_but_preserve_authorityless_relative_path() for detail.

pub fn and_normalize(self) -> Self

Returns Self with normalization enabled.

pub fn and_normalize_but_preserve_authorityless_relative_path(self) -> Self

Returns Self with special normalization enabled.

Note that this normalization algorithm is not compatible with RFC 3986 algorithm for some inputs.

See RiStr::normalize_but_preserve_authorityless_relative_path() for detail.

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Checks if the path is normalizable by RFC 3986 algorithm.

Returns Ok(()) when normalizable, returns Err(_) if not.