iri_string::percent_encode::decode

Struct PercentDecodedWhatwgBytes

pub struct PercentDecodedWhatwgBytes<'a> { /* private fields */ }

A percent-decoded string based on percent-decode algorithm of the WHATWG URL standard.

Note that this type does not guarantee that the string is valid percent-encoded string. The raw string may have stray percent character or the following digits may be invalid as hexadecimal digits.

Note that comparisons and hashing via std traits (such as Eq and Hash) will use the raw value, not the content after decoding.

Implementations

impl<'a> PercentDecodedWhatwgBytes<'a>

pub fn not_yet_decoded(&self) -> &'a [u8]

Returns the not-yet-decoded input bytes.

Examples

use iri_string::percent_encode::decode::decode_whatwg_bytes;

let decoded = decode_whatwg_bytes(b"hello%20world");

assert_eq!(decoded.not_yet_decoded(), &b"hello%20world"[..]);

pub fn to_bytes(&self) -> Option<&'a [u8]>

Returns the decoded bytes as a slice if no memory allocation is needed.

Examples

use iri_string::percent_encode::decode::decode_whatwg_bytes;

let no_alloc = decode_whatwg_bytes(b"99% unsafe");
assert_eq!(no_alloc.to_bytes(), Some(&b"99% unsafe"[..]));

let alloc_needed = decode_whatwg_bytes(b"hello%20world");
assert_eq!(alloc_needed.to_bytes(), None);

pub fn into_bytes(&self) -> Cow<'a, [u8]>

Decodes the bytes, based on percent-decode algorithm of the WHATWG URL standard.

Examples

use iri_string::percent_encode::decode::decode_whatwg_bytes;

let decoded = decode_whatwg_bytes(b"hello%20world");
assert_eq!(decoded.into_bytes(), &b"hello world"[..]);

pub fn into_string(&self) -> Result<String, FromUtf8Error>

Decodes the bytes into a string, based on percent-decode algorithm of the WHATWG URL standard.

Examples

use iri_string::percent_encode::decode::decode_whatwg_bytes;

let decoded = decode_whatwg_bytes(b"hello%20world");
assert_eq!(
    decoded.into_string(),
    Ok("hello world".to_owned())
);

pub fn bytes_fragments(&self) -> PercentDecodedBytesFragments<'a>

Returns an iterator of decoded fragments.

Examples

use iri_string::percent_encode::decode::{
    decode_whatwg_bytes, DecodedFragment,
};

let mut i = decode_whatwg_bytes(b"100% hello%20world")
    .bytes_fragments();

assert_eq!(i.next(), Some(DecodedFragment::Direct(b"100")));
assert_eq!(i.next(), Some(DecodedFragment::StrayPercent));
assert_eq!(i.next(), Some(DecodedFragment::Direct(b" hello")));
assert_eq!(i.next(), Some(DecodedFragment::DecodedByte(b' ')));
assert_eq!(i.next(), Some(DecodedFragment::Direct(b"world")));
assert_eq!(i.next(), None);