iri_string::template::context

Trait VisitValueNaive

pub trait VisitValueNaive {
    // Required method
    fn visit_value<V: Visitor>(&self, visitor: V) -> V::Result;
}

A trait for types that are expected to have “naive” visit implementation.

This trait can be implemented when a type of a value has a default variable type in template expansion context, and the value (or its Rust type) itself knows the type in template expansion context.

Required Methods

fn visit_value<V: Visitor>(&self, visitor: V) -> V::Result

Make visitor visit the value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.