Struct iri_string::template::UriTemplateString

pub struct UriTemplateString { /* private fields */ }

An owned slice of a URI template.

URI Template is defined by RFC 6570.

Note that “URI Template” can also be used for IRI.

Valid values

This type can have a URI template string.

Implementations

impl UriTemplateString

pub unsafe fn new_unchecked(s: String) -> Self

Creates a new string without validation.

This does not validate the given string, so it is caller’s responsibility to ensure the given string is valid.

Safety

The given string must be syntactically valid as Self type. If not, any use of the returned value or the call of this function itself may result in undefined behavior.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the inner buffer to match its length.

pub fn capacity(&self) -> usize

Returns the internal buffer capacity in bytes.

pub fn as_slice(&self) -> &UriTemplateStr

Returns the borrowed IRI string slice.

This is equivalent to &*self.

pub fn append(&mut self, other: &UriTemplateStr)

Appends the template string.