Struct iri_string::types::RiAbsoluteStr

pub struct RiAbsoluteStr<S> { /* private fields */ }

A borrowed slice of an absolute IRI without fragment part.

This corresponds to absolute-IRI rule in RFC 3987 (and absolute-URI rule in RFC 3986). In other words, this is RiStr without fragment part.

If you want to accept fragment part, use RiStr.

Valid values

This type can have an absolute IRI without fragment part.

assert!(IriAbsoluteStr::new("https://example.com/foo?bar=baz").is_ok());
assert!(IriAbsoluteStr::new("foo:bar").is_ok());
// Scheme `foo` and empty path.
assert!(IriAbsoluteStr::new("foo:").is_ok());
// `foo://.../` below are all allowed. See the crate documentation for detail.
assert!(IriAbsoluteStr::new("foo:/").is_ok());
assert!(IriAbsoluteStr::new("foo://").is_ok());
assert!(IriAbsoluteStr::new("foo:///").is_ok());
assert!(IriAbsoluteStr::new("foo:////").is_ok());
assert!(IriAbsoluteStr::new("foo://///").is_ok());

Relative IRI is not allowed.

// This is relative path.
assert!(IriAbsoluteStr::new("foo/bar").is_err());
// `/foo/bar` is an absolute path, but it is authority-relative.
assert!(IriAbsoluteStr::new("/foo/bar").is_err());
// `//foo/bar` is termed "network-path reference",
// or usually called "protocol-relative reference".
assert!(IriAbsoluteStr::new("//foo/bar").is_err());
// Empty string is not a valid absolute IRI.
assert!(IriAbsoluteStr::new("").is_err());

Fragment part (such as trailing #foo) is not allowed.

// Fragment part is not allowed.
assert!(IriAbsoluteStr::new("https://example.com/foo?bar=baz#qux").is_err());

Some characters and sequences cannot used in an absolute IRI.

// `<` and `>` cannot directly appear in an absolute IRI.
assert!(IriAbsoluteStr::new("<not allowed>").is_err());
// Broken percent encoding cannot appear in an absolute IRI.
assert!(IriAbsoluteStr::new("%").is_err());
assert!(IriAbsoluteStr::new("%GG").is_err());

Implementations

impl<S: Spec> RiAbsoluteStr<S>

pub fn new(s: &str) -> Result<&Self, Error>

Creates a new string.

pub unsafe fn new_unchecked(s: &str) -> &Self

Creates a new string without validation.

This does not validate the given string, so it is caller’s responsibility to ensure the given string is valid.

Safety

The given string must be syntactically valid as Self type. If not, any use of the returned value or the call of this function itself may result in undefined behavior.

pub fn as_str(&self) -> &str

Returns &str.

pub fn len(&self) -> usize

Returns the string length.

pub fn is_empty(&self) -> bool

Returns whether the string is empty.

impl<S: Spec> RiAbsoluteStr<S>

pub fn ensure_rfc3986_normalizable(&self) -> Result<(), Error>

Returns Ok(()) if the IRI is normalizable by the RFC 3986 algorithm.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/%2e/bar/..")?;
assert!(iri.ensure_rfc3986_normalizable().is_ok());

let iri2 = IriAbsoluteStr::new("scheme:/..//bar")?;
// The normalization result would be `scheme://bar` according to RFC
// 3986, but it is unintended and should be treated as a failure.
// This crate automatically handles this case so that `.normalize()` won't fail.
assert!(!iri.ensure_rfc3986_normalizable().is_err());

pub fn is_normalized(&self) -> bool

Returns true if the IRI is already normalized.

This returns the same result as self.normalize().to_string() == self, but does this more efficiently without heap allocation.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/./bar/%2e%2e/../baz?query")?;
assert!(!iri.is_normalized());

let normalized = iri.normalize().to_dedicated_string();
assert_eq!(normalized, "http://example.com/baz?query");
assert!(normalized.is_normalized());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:/.///foo")?;
// Already normalized.
assert!(iri.is_normalized());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:relative/..//not-a-host")?;
// Default normalization algorithm assumes the path part to be NOT opaque.
assert!(!iri.is_normalized());

let normalized = iri.normalize().to_dedicated_string();
assert_eq!(normalized, "scheme:/.//not-a-host");

pub fn is_normalized_rfc3986(&self) -> bool

Returns true if the IRI is already normalized.

This returns the same result as self.ensure_rfc3986_normalizable() && (self.normalize().to_string() == self), does this more efficiently without heap allocation.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/./bar/%2e%2e/../baz?query")?;
assert!(!iri.is_normalized_rfc3986());

let normalized = iri.normalize().to_dedicated_string();
assert_eq!(normalized, "http://example.com/baz?query");
assert!(normalized.is_normalized_rfc3986());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:/.///foo")?;
// Not normalized in the sense of RFC 3986.
assert!(!iri.is_normalized_rfc3986());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:relative/..//not-a-host")?;
// RFC 3986 normalization algorithm assumes the path part to be NOT opaque.
assert!(!iri.is_normalized_rfc3986());

let normalized = iri.normalize().to_dedicated_string();
assert_eq!(normalized, "scheme:/.//not-a-host");

pub fn is_normalized_but_authorityless_relative_path_preserved(&self) -> bool

Returns true if the IRI is already normalized in the sense of normalize_but_preserve_authorityless_relative_path method.

This returns the same result as self.normalize_but_preserve_authorityless_relative_path().to_string() == self, but does this more efficiently without heap allocation.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/./bar/%2e%2e/../baz?query")?;
assert!(!iri.is_normalized_but_authorityless_relative_path_preserved());

let normalized = iri
    .normalize_but_preserve_authorityless_relative_path()
    .to_dedicated_string();
assert_eq!(normalized, "http://example.com/baz?query");
assert!(normalized.is_normalized());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:/.///foo")?;
// Already normalized in the sense of
// `normalize_but_opaque_authorityless_relative_path()` method.
assert!(iri.is_normalized_but_authorityless_relative_path_preserved());

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:relative/..//not-a-host")?;
// Relative path is treated as opaque since the autority component is absent.
assert!(iri.is_normalized_but_authorityless_relative_path_preserved());

pub fn normalize(&self) -> Normalized<'_, Self>

Returns the normalized IRI.

Notes

For some abnormal IRIs, the normalization can produce semantically incorrect string that looks syntactically valid. To avoid security issues by this trap, the normalization algorithm by this crate automatically applies the workaround.

If you worry about this, test by RiAbsoluteStr::ensure_rfc3986_normalizable method or Normalized::ensure_rfc3986_normalizable before using the result string.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/./bar/%2e%2e/../baz?query")?;

let normalized = iri.normalize().to_dedicated_string();
assert_eq!(normalized, "http://example.com/baz?query");

pub fn normalize_but_preserve_authorityless_relative_path( &self, ) -> Normalized<'_, Self>

Returns the normalized IRI, but preserving dot segments in relative path if the authority component is absent.

This normalization would be similar to that of WHATWG URL Standard while this implementation is not guaranteed to stricly follow the spec.

Note that this normalization algorithm is not compatible with RFC 3986 algorithm for some inputs.

Note that case normalization and percent-encoding normalization will still be applied to any path.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("HTTP://example.COM/foo/./bar/%2e%2e/../baz?query")?;

let normalized = iri
    .normalize_but_preserve_authorityless_relative_path()
    .to_dedicated_string();
assert_eq!(normalized, "http://example.com/baz?query");

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("scheme:relative/../f%6f%6f")?;

let normalized = iri
    .normalize_but_preserve_authorityless_relative_path()
    .to_dedicated_string();
assert_eq!(normalized, "scheme:relative/../foo");
// `.normalize()` would normalize this to `scheme:/foo`.

pub fn mask_password(&self) -> PasswordMasked<'_, Self>

Returns the proxy to the IRI with password masking feature.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();
assert_eq!(masked.to_dedicated_string(), "http://user:@example.com/path?query");

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

impl<S: Spec> RiAbsoluteStr<S>

Components getters.

pub fn scheme_str(&self) -> &str

Returns the scheme.

The following colon is truncated.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://example.com/pathpath?queryquery")?;
assert_eq!(iri.scheme_str(), "http");

pub fn authority_str(&self) -> Option<&str>

Returns the authority.

The leading // is truncated.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://example.com/pathpath?queryquery")?;
assert_eq!(iri.authority_str(), Some("example.com"));

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.authority_str(), None);

pub fn path_str(&self) -> &str

Returns the path.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://example.com/pathpath?queryquery")?;
assert_eq!(iri.path_str(), "/pathpath");

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.path_str(), "uuid:10db315b-fcd1-4428-aca8-15babc9a2da2");

pub fn query(&self) -> Option<&RiQueryStr<S>>

Returns the query.

The leading question mark (?) is truncated.

Examples

use iri_string::types::{IriAbsoluteStr, IriQueryStr};

let iri = IriAbsoluteStr::new("http://example.com/pathpath?queryquery")?;
let query = IriQueryStr::new("queryquery")?;
assert_eq!(iri.query(), Some(query));

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.query(), None);

pub fn query_str(&self) -> Option<&str>

Returns the query in a raw string slice.

The leading question mark (?) is truncated.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://example.com/pathpath?queryquery")?;
assert_eq!(iri.query_str(), Some("queryquery"));

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.query_str(), None);

pub fn authority_components(&self) -> Option<AuthorityComponents<'_>>

Returns the authority components.

Examples

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("http://user:pass@example.com:8080/pathpath?queryquery")?;
let authority = iri.authority_components()
    .expect("authority is available");
assert_eq!(authority.userinfo(), Some("user:pass"));
assert_eq!(authority.host(), "example.com");
assert_eq!(authority.port(), Some("8080"));

use iri_string::types::IriAbsoluteStr;

let iri = IriAbsoluteStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.authority_str(), None);

impl RiAbsoluteStr<IriSpec>

Conversion from an IRI into a URI.

pub fn encode_to_uri(&self) -> MappedToUri<'_, Self>

Percent-encodes the IRI into a valid URI that identifies the equivalent resource.

If you need more precise control over memory allocation and buffer handling, use MappedToUri type.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::{IriAbsoluteStr, UriAbsoluteString};

let iri = IriAbsoluteStr::new("http://example.com/?alpha=\u{03B1}")?;
// Type annotation here is not necessary.
let uri: UriAbsoluteString = iri.encode_to_uri().to_dedicated_string();
assert_eq!(uri, "http://example.com/?alpha=%CE%B1");

pub fn as_uri(&self) -> Option<&UriAbsoluteStr>

Converts an IRI into a URI without modification, if possible.

This is semantically equivalent to UriAbsoluteStr::new(self.as_str()).ok().

Examples

use iri_string::types::{IriAbsoluteStr, UriAbsoluteStr};

let ascii_iri = IriAbsoluteStr::new("http://example.com/?alpha=%CE%B1")?;
assert_eq!(
    ascii_iri.as_uri().map(AsRef::as_ref),
    Some("http://example.com/?alpha=%CE%B1")
);

let nonascii_iri = IriAbsoluteStr::new("http://example.com/?alpha=\u{03B1}")?;
assert_eq!(nonascii_iri.as_uri(), None);