Struct iri_string::types::RiReferenceStr

pub struct RiReferenceStr<S> { /* private fields */ }

A borrowed string of an absolute IRI possibly with fragment part.

This corresponds to IRI-reference rule in RFC 3987 (and URI-reference rule in RFC 3986). The rule for IRI-reference is IRI / irelative-ref. In other words, this is union of RiStr and RiRelativeStr.

Valid values

This type can have an IRI reference (which can be absolute or relative).

assert!(IriReferenceStr::new("https://user:pass@example.com:8080").is_ok());
assert!(IriReferenceStr::new("https://example.com/").is_ok());
assert!(IriReferenceStr::new("https://example.com/foo?bar=baz").is_ok());
assert!(IriReferenceStr::new("https://example.com/foo?bar=baz#qux").is_ok());
assert!(IriReferenceStr::new("foo:bar").is_ok());
assert!(IriReferenceStr::new("foo:").is_ok());
// `foo://.../` below are all allowed. See the crate documentation for detail.
assert!(IriReferenceStr::new("foo:/").is_ok());
assert!(IriReferenceStr::new("foo://").is_ok());
assert!(IriReferenceStr::new("foo:///").is_ok());
assert!(IriReferenceStr::new("foo:////").is_ok());
assert!(IriReferenceStr::new("foo://///").is_ok());
assert!(IriReferenceStr::new("foo/bar").is_ok());
assert!(IriReferenceStr::new("/foo/bar").is_ok());
assert!(IriReferenceStr::new("//foo/bar").is_ok());
assert!(IriReferenceStr::new("#foo").is_ok());

Some characters and sequences cannot used in an IRI reference.

// `<` and `>` cannot directly appear in an IRI reference.
assert!(IriReferenceStr::new("<not allowed>").is_err());
// Broken percent encoding cannot appear in an IRI reference.
assert!(IriReferenceStr::new("%").is_err());
assert!(IriReferenceStr::new("%GG").is_err());

Implementations

impl<S: Spec> RiReferenceStr<S>

pub fn new(s: &str) -> Result<&Self, Error>

Creates a new string.

pub unsafe fn new_unchecked(s: &str) -> &Self

Creates a new string without validation.

This does not validate the given string, so it is caller’s responsibility to ensure the given string is valid.

Safety

The given string must be syntactically valid as Self type. If not, any use of the returned value or the call of this function itself may result in undefined behavior.

pub fn as_str(&self) -> &str

Returns &str.

pub fn len(&self) -> usize

Returns the string length.

pub fn is_empty(&self) -> bool

Returns whether the string is empty.

impl<S: Spec> RiReferenceStr<S>

pub fn to_iri(&self) -> Result<&RiStr<S>, &RiRelativeStr<S>>

Returns the string as &RiStr, if it is valid as an IRI.

If it is not an IRI, then &RiRelativeStr is returned as Err(_).

pub fn to_relative_iri(&self) -> Result<&RiRelativeStr<S>, &RiStr<S>>

Returns the string as &RiRelativeStr, if it is valid as an IRI.

If it is not an IRI, then &RiStr is returned as Err(_).

pub fn resolve_against<'a>( &'a self, base: &'a RiAbsoluteStr<S>, ) -> Normalized<'a, RiStr<S>>

Returns resolved IRI against the given base IRI.

For IRI reference resolution output examples, see RFC 3986 section 5.4.

If you are going to resolve multiple references against the common base, consider using FixedBaseResolver.

Strictness

The IRI parsers provided by this crate is strict (e.g. http:g is always interpreted as a composition of the scheme http and the path g), so backward compatible parsing and resolution are not provided. About parser and resolver strictness, see RFC 3986 section 5.4.2:

Some parsers allow the scheme name to be present in a relative reference if it is the same as the base URI scheme. This is considered to be a loophole in prior specifications of partial URI RFC1630. Its use should be avoided but is allowed for backward compatibility.

— https://tools.ietf.org/html/rfc3986#section-5.4.2

Failures

This method itself does not fail, but IRI resolution without WHATWG URL Standard serialization can fail in some minor cases.

To see examples of such unresolvable IRIs, visit the documentation for normalize module.

pub fn mask_password(&self) -> PasswordMasked<'_, Self>

Returns the proxy to the IRI with password masking feature.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:password@example.com/path?query")?;
let masked = iri.mask_password();
assert_eq!(masked.to_dedicated_string(), "http://user:@example.com/path?query");

assert_eq!(
    masked.replace_password("${password}").to_string(),
    "http://user:${password}@example.com/path?query"
);

impl<S: Spec> RiReferenceStr<S>

Components getters.

pub fn scheme_str(&self) -> Option<&str>

Returns the scheme.

The following colon is truncated.

Examples

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://example.com/pathpath?queryquery#fragfrag")?;
assert_eq!(iri.scheme_str(), Some("http"));

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("foo/bar:baz")?;
assert_eq!(iri.scheme_str(), None);

pub fn authority_str(&self) -> Option<&str>

Returns the authority.

The leading // is truncated.

Examples

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://example.com/pathpath?queryquery#fragfrag")?;
assert_eq!(iri.authority_str(), Some("example.com"));

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.authority_str(), None);

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("foo/bar:baz")?;
assert_eq!(iri.authority_str(), None);

pub fn path_str(&self) -> &str

Returns the path.

Examples

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://example.com/pathpath?queryquery#fragfrag")?;
assert_eq!(iri.path_str(), "/pathpath");

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.path_str(), "uuid:10db315b-fcd1-4428-aca8-15babc9a2da2");

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("foo/bar:baz")?;
assert_eq!(iri.path_str(), "foo/bar:baz");

pub fn query(&self) -> Option<&RiQueryStr<S>>

Returns the query.

The leading question mark (?) is truncated.

Examples

use iri_string::types::{IriQueryStr, IriReferenceStr};

let iri = IriReferenceStr::new("http://example.com/pathpath?queryquery#fragfrag")?;
let query = IriQueryStr::new("queryquery")?;
assert_eq!(iri.query(), Some(query));

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.query(), None);

use iri_string::types::{IriQueryStr, IriReferenceStr};

let iri = IriReferenceStr::new("foo/bar:baz?")?;
let query = IriQueryStr::new("")?;
assert_eq!(iri.query(), Some(query));

pub fn query_str(&self) -> Option<&str>

Returns the query as a raw string slice.

The leading question mark (?) is truncated.

Examples

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://example.com/pathpath?queryquery#fragfrag")?;
assert_eq!(iri.query_str(), Some("queryquery"));

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
assert_eq!(iri.query_str(), None);

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("foo/bar:baz?")?;
assert_eq!(iri.query_str(), Some(""));

pub fn fragment(&self) -> Option<&RiFragmentStr<S>>

Returns the fragment part if exists.

A leading # character is truncated if the fragment part exists.

Examples

If the IRI has a fragment part, Some(_) is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux#corge")?;
let fragment = IriFragmentStr::new("corge")?;
assert_eq!(iri.fragment(), Some(fragment));

let iri = IriReferenceStr::new("#foo")?;
let fragment = IriFragmentStr::new("foo")?;
assert_eq!(iri.fragment(), Some(fragment));

When the fragment part exists but is empty string, Some(_) is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux#")?;
let fragment = IriFragmentStr::new("")?;
assert_eq!(iri.fragment(), Some(fragment));

let iri = IriReferenceStr::new("#")?;
let fragment = IriFragmentStr::new("")?;
assert_eq!(iri.fragment(), Some(fragment));

If the IRI has no fragment, None is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux")?;
assert_eq!(iri.fragment(), None);

let iri = IriReferenceStr::new("")?;
assert_eq!(iri.fragment(), None);

pub fn fragment_str(&self) -> Option<&str>

Returns the fragment part as a raw string slice if exists.

A leading # character is truncated if the fragment part exists.

Examples

If the IRI has a fragment part, Some(_) is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux#corge")?;
assert_eq!(iri.fragment_str(), Some("corge"));

let iri = IriReferenceStr::new("#foo")?;
assert_eq!(iri.fragment_str(), Some("foo"));

When the fragment part exists but is empty string, Some(_) is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux#")?;
assert_eq!(iri.fragment_str(), Some(""));

let iri = IriReferenceStr::new("#")?;
assert_eq!(iri.fragment_str(), Some(""));

If the IRI has no fragment, None is returned.

let iri = IriReferenceStr::new("foo://bar/baz?qux=quux")?;
assert_eq!(iri.fragment(), None);

let iri = IriReferenceStr::new("")?;
assert_eq!(iri.fragment(), None);

pub fn authority_components(&self) -> Option<AuthorityComponents<'_>>

Returns the authority components.

Examples

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("http://user:pass@example.com:8080/pathpath?queryquery")?;
let authority = iri.authority_components()
    .expect("authority is available");
assert_eq!(authority.userinfo(), Some("user:pass"));
assert_eq!(authority.host(), "example.com");
assert_eq!(authority.port(), Some("8080"));

use iri_string::types::IriReferenceStr;

let iri = IriReferenceStr::new("foo//bar:baz")?;
assert_eq!(iri.authority_str(), None);

impl RiReferenceStr<IriSpec>

Conversion from an IRI into a URI.

pub fn encode_to_uri(&self) -> MappedToUri<'_, Self>

Percent-encodes the IRI into a valid URI that identifies the equivalent resource.

If you need more precise control over memory allocation and buffer handling, use MappedToUri type.

Examples

use iri_string::format::ToDedicatedString;
use iri_string::types::{IriReferenceStr, UriReferenceString};

let iri = IriReferenceStr::new("http://example.com/?alpha=\u{03B1}")?;
// Type annotation here is not necessary.
let uri: UriReferenceString = iri.encode_to_uri().to_dedicated_string();
assert_eq!(uri, "http://example.com/?alpha=%CE%B1");

pub fn as_uri(&self) -> Option<&UriReferenceStr>

Converts an IRI into a URI without modification, if possible.

This is semantically equivalent to UriReferenceStr::new(self.as_str()).ok().

Examples

use iri_string::types::{IriReferenceStr, UriReferenceStr};

let ascii_iri = IriReferenceStr::new("http://example.com/?alpha=%CE%B1")?;
assert_eq!(
    ascii_iri.as_uri().map(AsRef::as_ref),
    Some("http://example.com/?alpha=%CE%B1")
);

let nonascii_iri = IriReferenceStr::new("http://example.com/?alpha=\u{03B1}")?;
assert_eq!(nonascii_iri.as_uri(), None);