iri_string

Module template

Processor for RFC 6570 URI Template.

Usage

Prepare a template.

You can create a template as UriTemplateStr type (borrowed) or UriTemplateString type (owned).

Prepare a context.

Create a value of type that implements Context trait.

Or, if you use SimpleContext, insert key-value pairs into it.

Expand.

Pass the context to UriTemplateStr::expand method of the template.

Use the result.

Returned Expanded object can be directly printed since it implements Display trait. Or, you can call .to_string() method of the alloc::string::ToString trait to convert it to a String.

Examples

Custom context type

For details, see the documentation of context module.

use core::fmt;
use iri_string::spec::{IriSpec, Spec, UriSpec};
use iri_string::template::UriTemplateStr;
use iri_string::template::context::{Context, VarName, Visitor};

struct UserInfo {
    username: &'static str,
    utf8_available: bool,
}

impl Context for UserInfo {
    fn visit<V: Visitor>(
        &self,
        visitor: V,
    ) -> V::Result {
        match visitor.var_name().as_str() {
            "username" => visitor.visit_string(self.username),
            "utf8" => {
                if self.utf8_available {
                    // U+2713 CHECK MARK
                    visitor.visit_string("\u{2713}")
                } else {
                    visitor.visit_undefined()
                }
            }
            _ => visitor.visit_undefined()
        }
    }
}

let context = UserInfo {
    username: "foo",
    utf8_available: true,
};

let template = UriTemplateStr::new("/users/{username}{?utf8}")?;

assert_eq!(
    template.expand::<UriSpec, _>(&context)?.to_string(),
    "/users/foo?utf8=%E2%9C%93"
);
assert_eq!(
    template.expand::<IriSpec, _>(&context)?.to_string(),
    "/users/foo?utf8=\u{2713}"
);

SimpleContext type (enabled by alloc feature flag)

use iri_string::spec::{IriSpec, UriSpec};
use iri_string::template::UriTemplateStr;
use iri_string::template::simple_context::SimpleContext;

let mut context = SimpleContext::new();
context.insert("username", "foo");
// U+2713 CHECK MARK
context.insert("utf8", "\u{2713}");

let template = UriTemplateStr::new("/users/{username}{?utf8}")?;

assert_eq!(
    template.expand::<UriSpec, _>(&context)?.to_string(),
    "/users/foo?utf8=%E2%9C%93"
);
assert_eq!(
    template.expand::<IriSpec, _>(&context)?.to_string(),
    "/users/foo?utf8=\u{2713}"
);

Re-exports

pub use self::context::Context;

pub use self::context::DynamicContext;

Modules

context

Template expansion context.

simple_context

Simple general-purpose context type.

Structs

CreationError

Error on conversion into a URI template type.

Error

Template construction and expansion error.

Expanded

Template expansion result.

UriTemplateStr

A borrowed slice of a URI template.

UriTemplateString

An owned slice of a URI template.

UriTemplateVariables

An iterator of variables in a URI template.

Type Aliases

VarNameDeprecated

Deprecated old name of template::context::VarName.